[server] Support projection pushdown with field id in server.

+++ fluss-server/src/main/java/org/apache/fluss/server/coordinator/SchemaUpdate.java
@@ -63,9 +63,7 @@ public class SchemaUpdate {
 
     public Schema getSchema() {
         Schema.Builder builder =
-                Schema.newBuilder()
-                        .fromColumns(columns)
-                        .highestFieldId((short) highestFieldId.get());
+                Schema.newBuilder().fromColumns(columns).highestFieldId(highestFieldId.get());
         if (!primaryKeys.isEmpty()) {
             builder.primaryKey(primaryKeys);
         }
